refactor: moved all model calls like retraining, loading, and saving into a model service file for ease of use.
